requirements.txt
Django==1.11.10
django-filter==1.1.0
django-rest-auth==0.9.3
djangorestframework==3.7.7
Markdown==2.6.11
pytz==2018.3
six==1.11.0
